v.0.4 Multiple delete.

+++ src/main/java/by/htp/ahremenko/controller/NewsController.java
@@ -1,27 +1,21 @@
 package by.htp.ahremenko.controller;
+import java.text.SimpleDateFormat;
+import java.util.Date;
 import java.util.List;
-
+import javax.servlet.http.HttpServletRequest;
 import javax.validation.Valid;
-
-//import javax.validation.Valid;
-
-//import org.hibernate.Session;
-//import org.hibernate.SessionFactory;
-//import org.hibernate.cfg.Configuration;
+import org.apache.log4j.Logger;
 import org.springframework.beans.factory.annotation.Autowired;
-//import org.springframework.beans.propertyeditors.StringTrimmerEditor;
+import org.springframework.beans.propertyeditors.CustomDateEditor;
 import org.springframework.stereotype.Controller;
 import org.springframework.ui.Model;
 import org.springframework.validation.BindingResult;
-import org.springframework.web.bind.annotation.ModelAttribute;
+import org.springframework.web.bind.WebDataBinder;
 //import org.springframework.validation.BindingResult;
-//import org.springframework.web.bind.WebDataBinder;
-//import org.springframework.web.bind.annotation.InitBinder;
-//import org.springframework.web.bind.annotation.ModelAttribute;
+import org.springframework.web.bind.annotation.ModelAttribute;
 import org.springframework.web.bind.annotation.RequestMapping;
-//import org.springframework.web.bind.annotation.RequestParam;
+import org.springframework.web.bind.annotation.RequestMethod;
 import org.springframework.web.bind.annotation.RequestParam;
-
 import by.htp.ahremenko.domain.News;
 import by.htp.ahremenko.service.NewsService;
 import by.htp.ahremenko.service.exception.ServiceException;
@@ -35,9 +29,12 @@ public class NewsController {
 	@Autowired
 	private NewsService newsService; 
 	
+	private static final Logger logger = Logger.getLogger(NewsController.class);
+	
 	@RequestMapping("/listNews")
 	public String processForm(Model theModel) {
 		
+		logger.info("Get list of news done.");
 		try {
 			List<News> result = newsService.getList(""); 
 			theModel.addAttribute("newsList", result);		
@@ -53,7 +50,6 @@ public class NewsController {
 	public String showForm(Model theModel) {
 		News theNews= new News();
 		theModel.addAttribute("news", theNews); 
-		//System.out.println("New User was put into model. [" + theUser.getId() +"]");
 		return "edit-form";
 	}
 	
@@ -67,44 +63,60 @@ public class NewsController {
 			theModel.addAttribute("errorMessage", "Error: " + e.getMessage());
 			return "error";
 		}	
-		
-		//System.out.println("New User was put into model. [" + theUser.getId() +"]");
-		
 	}
 	
 	@RequestMapping("/saveNews")
-	public String processForm(@Valid @ModelAttribute("news") News theNews, Model theModel) {
-
-		try {
-			newsService.save(theNews);
-			List<News> result = newsService.getList(""); 
-			theModel.addAttribute("newsList", result);		
-			return "newsmanagement";
-		} catch (ServiceException e){
-			theModel.addAttribute("errorMessage", "Error: " + e.getMessage());
-			return "error";
-		}
+	public String processForm(Model theModel, @Valid @ModelAttribute("news") News theNews, BindingResult theBindingResult) {
 		
+		logger.info("News ID: " + theNews.getId() + " will be saved.");
+		if (theBindingResult.hasErrors()) {
+			return "edit-form";
+		} else {
+			try {
+				newsService.save(theNews);
+				List<News> result = newsService.getList(""); 
+				theModel.addAttribute("newsList", result);		
+				return "newsmanagement";
+			} catch (ServiceException e){
+				theModel.addAttribute("errorMessage", "Error: " + e.getMessage());
+				return "error";
+			}
+		}
 	}
 	
-	@RequestMapping("/delete")
-	public String deleteNews(@RequestParam("newsId") int theId, Model theModel) {
+	@RequestMapping(path = "/delete", method = RequestMethod.POST)
+	public String deleteNews(Model theModel, HttpServletRequest request) {
 		
-		System.out.println("Now ID: " + theId + " will be deleted.");
-		try {
-			newsService.remove(theId);
-			List<News> result = newsService.getList(""); 
-			theModel.addAttribute("newsList", result);		
-			return "newsmanagement";
-		} catch (ServiceException e){
-			theModel.addAttribute("errorMessage", "Error: " + e.getMessage());
-			return "error";
-		}
-		
+		if ( request.getParameterValues("selectedIds") != null ) {
+			try {
+				
+				for (String theIds : request.getParameterValues("selectedIds")) {
+					int theId = Integer.parseInt(theIds);
+					newsService.remove(theId);
+					//logger.info("Now News ID: " + theIds + " will be deleted!");
+				}
+				
+				List<News> result = newsService.getList(""); 
+				theModel.addAttribute("newsList", result);		
+				
+			} catch (ServiceException e){
+				theModel.addAttribute("errorMessage", "Error: " + e.getMessage());
+				return "error";
+			}	
+				
+		}	
+		return "newsmanagement";		
 	}
+
 	
 	@RequestMapping("/about")	
 	public String goToAboutPage() {
 		return "about";
 	}
+	
+	/*public void initBinder(WebDataBinder binder){
+        SimpleDateFormat sdf = new SimpleDateFormat("MM.dd.yyyy");
+        sdf.setLenient(true);
+        binder.registerCustomEditor(Date.class, new CustomDateEditor(sdf, true));
+    }*/
 }
